extractService/requirements.txt
uvicorn
fastapi
boto3
google-cloud-pubsub
pymupdf
pillow
pytesseract
